// Node_Whatsapp_Api/src/Interfaces/MessageGroupAPI.ts
export interface MessageGroupAPI {
    mediacontenttype: string | null;
    smsmessagesid: string;
    nummedia: number;
    smssid: string;
    body: string;
    src: string;
    accountsid: string;
    dst: string;
    mediaurl: string | null;
    date: string;
    cost_cli: string | null;
    latitude: number | null;
    longitude: number | null;
    profilename: string | null;
    notifyname: string | null;
    author: string | null;
    isforwarded: boolean;
    forwardingscore: number | null;
}